refactor: extract shared defineLazy into @confect/core Lazy.defineProperty
defineLazy was duplicated verbatim in @confect/core and @confect/server's
internal/utils, risking drift. Move the single implementation to a new
@confect/core `Lazy` module as `Lazy.defineProperty` and consume it from both
FunctionProvenance (core) and Table (server, via @confect/core/Lazy). Delete
core's internal/utils.ts (it held only this helper), drop defineLazy from
server's internal/utils, and move its unit tests to core/test/Lazy.test.ts.
